requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.7

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2